% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDadj_utility.R
\name{full_normal_geno}
\alias{full_normal_geno}
\title{Standardizing the genotype data.}
\usage{
full_normal_geno(geno_raw, max_size = 1e+05, NAval = NA)
}
\arguments{
\item{geno_raw}{the genotype matrix assuming each row is the individual and each column is the SNP.}

\item{max_size}{an integer for the maximum size of SNPs to be standardized at a time, the default is 100000.
This option can be ignored for data with less than 1 million SNPs.}

\item{NAval}{the missing value used in the output data matrix. The default is NA, for use with
PLINK, set \code{NAval} = -9.}
}
\value{
a data matrix of standardized genotypes.
}
\description{
The function standardizes a genotype data such that for each SNP the mean is 0 and standard
   deviation is 1.
}
\examples{
data(geno)
norm_geno <- full_normal_geno(geno_raw = geno)

}
