% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraB_utility.R
\name{get_formulas}
\alias{get_formulas}
\title{Regression Tree Formula}
\usage{
get_formulas(name, var_names)
}
\arguments{
\item{name}{a character to indicate the regression coefficients from which complex trait.
It should match the trait name used in the other analyses.}

\item{var_names}{the names of the variables to be included.}
}
\value{
a formula specifying the response variable (univariate regression coefficients from target population)
   and the predictors are used to tune the weights in boosted regression trees model.
}
\description{
The function generates the formula to be used in regression trees model.
}
\examples{
data(annotation_data)
var_names <- colnames(annotation_data)
formulas = get_formulas(name = 'BMI', var_names = var_names[2:5])
print(formulas)

}
