% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraB_utility.R
\name{get_NAs}
\alias{get_NAs}
\title{Remove missing values}
\usage{
get_NAs(data, NA_value = "NaN")
}
\arguments{
\item{data}{a numeric data matrix}

\item{NA_value}{the missing value code in the data, the default is \code{NaN}.
Could also be \code{NA} or \code{-9} commonly used in PLINK data.}
}
\value{
a data matrix of genotypes.
}
\description{
The function records and removes the missing values from the input data matrix.
}
