% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraB.R
\name{GraBLD.score}
\alias{GraBLD.score}
\title{Gradient Boosted and LD adjusted Prediction.}
\usage{
GraBLD.score(
  source_data = NULL,
  SNP_names = NULL,
  chr = NULL,
  geno_raw,
  PLINK = TRUE,
  max_size = 1e+05,
  NAval = NA,
  Pheno = NULL,
  LDadjVal,
  gbmVal,
  trait_name = NULL,
  WRITE = FALSE
)
}
\arguments{
\item{source_data}{the name of the file which the genotype data are to be read from.
Each row of the matrix appears as one line of the file. Could be an absolute path to the
file or the name of the file assuming in the present directory \code{\link{getwd}()}.}

\item{SNP_names}{a vector of characters for the names of SNPs matching the corresponding genotype data, these are used in the output of GraBLD weights.}

\item{chr}{an integer indicating the maximum number of chromosomes to be read in, this option is used in combination with
\code{source_data}, to perform analysis by each chromosome. In this case, the file name should follow:
``source_data_i.raw'' for all \eqn{i \le chr}. For example, \code{Geno_Matrix_23.raw}.}

\item{geno_raw}{the genotype matrix assuming each row is the individual and each column is the SNP. Can be skipped if \code{source_data} was provided.}

\item{PLINK}{a logic indicating whether the supplied file is of the .raw format from PLINK, if not, the first six
columns will not be removed and all columns will be read in.}

\item{max_size}{an integer for the maximum size of SNPs to be standardized at a time, the default is 100000.
This option can be ignored for data with less than 1 million SNPs.}

\item{NAval}{the missing value used in the output data matrix. The default is NA, for use with PLINK, set \code{NAval} = -9.}

\item{Pheno}{a numeric vector of quantitative traits with the same length as the number of rows in the genotype matrix.}

\item{LDadjVal}{a data.frame of LD adjusted weights. This can be taken from the output of \code{\link{LDadj}()}. The column names must be "SNP" and "LDadj".}

\item{gbmVal}{a data.frame of gradient boosted weights. This can be taken from the output of \code{\link{GraB}()}. The column names must be "SNP" and "GraB".}

\item{trait_name}{a character indicating the name of the quantitative trait.}

\item{WRITE}{a logic indicating whether the results of the GraBLD weights should be written to a file with file name \code{trait_name_gbm_beta.txt}.}
}
\value{
if \code{Pheno} is supplied, both the polygenic gene score as well as the prediction R-squared (adjusted) are returned, otherwise only the polygenic gene score is returned.
}
\description{
The function returns the prediction r-squared on the target population using polygenic gene score based on the GraBLD heuristic.
}
\examples{
data(geno)
data(univariate_beta)
data(annotation_data)
LD_val <- LDadj(geno_raw = geno[grepl("rs", names(geno))],
SNP_names = names(geno)[grepl("rs", names(geno))],
chr = 1, size = 200)
gbm_val <- list()
for (j in 1:5){
gbm_val[[j]] <- GraB(betas = univariate_beta,
annotations = annotation_data, trait_name = 'BMI',
steps = j, validation = 5)
}
data(BMI)
gs <- GraBLD.score(geno_raw = geno[grepl("rs", names(geno))],
 SNP_names = names(geno)[grepl("rs", names(geno))],
 LDadjVal=LD_val, gbmVal = do.call(rbind, gbm_val),
 trait_name='BMI', Pheno = BMI[,1])

}
\references{
Paré, G., Mao, S. & Deng, W.Q. A machine-learning heuristic to improve gene score prediction of polygenic traits. \emph{Scientific reports} \strong{7}, 12665 (2017). \doi{10.1038/s41598-017-13056-1}.
}
