% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDadj_utility.R
\name{regionalLD}
\alias{regionalLD}
\title{Regional LD adjustment}
\usage{
regionalLD(geno_data, position = "start", size = 300)
}
\arguments{
\item{geno_data}{the standardized genotype matrix assuming each row is the individual and each column is the SNP.}

\item{position}{a character indicating the position of the SNP in a block of SNPs, one of
\code{start} (the start of the block), \code{mid} (the middle of the block),
\code{end} (the end of the block), or \code{together} (combining SNPs in a few blocks),
\code{short} (in a short block), or \code{all} (using all regions).}

\item{size}{an integer for the number of SNPs that should be included in a block. Usually, for genome-wide
datasets, there are about 2 million SNPs and +/- 300 SNPs is roughly equivalent to 1Mb
physical distance, thus 300 is set as the default size.}
}
\value{
a numeric vector of LD adjustments with the same length as the number of SNPs in the genotype matrix.
}
\description{
The function calculates the LD adjustment of each SNP for one given size of
   genotype data block.
}
\examples{
data(geno)
norm_geno <- full_normal_geno(geno_raw = geno)
regLD <- regionalLD(geno_data = norm_geno, position = 'mid', size = 300)
print(regLD)

}
\references{
Pare, Guillaume, Shihong Mao, and Wei Q. Deng.
A method to estimate the contribution of regional genetic
associations to complex traits from summary association statistics.
\emph{Scientific reports} \strong{6} (2016): 27644.
}
