% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraB_utility.R
\name{load_beta}
\alias{load_beta}
\title{Load univariate regression beta coefficients.}
\usage{
load_beta(trait_name, file_name = NULL)
}
\arguments{
\item{trait_name}{a character for the name of the quantitative trait, assuming the file is named as \code{trait_name_univariate_beta.txt}.}

\item{file_name}{a character to specify the full directory of the univariate regression coefficient file instead of \code{trait_name_univariate_beta.txt}.}
}
\value{
a data matrix of univariate regression coefficients.
}
\description{
The function loads the univariate regression beta coefficients into the workspace.
}
\details{
The univariate regression coefficients are assumed to be computed univariately using: \cr
   \code{coef(summary(lm(pheno_data ~ geno[,j])))[2,1]}.
   Both genotype data and phenotype data over individuals need to be standardized with
   mean = 0 and variance = 1. The data file should have two columns indicating chromosome number
   and the regression coefficients without headers.
}
